+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - #341 - Split connection delegator into batch delegator and single_delegator
 - #351 - Rename `#retrieve!` to `#parse!` on params and `#parsed` to `parse!` on params batch.
 - #351 - Adds '#first' for params_batch that returns parsed first element from the params_batch object.
+- #359 - Divide mark_as_consumed into mark_as_consumed and mark_as_consumed!
 - #356 - Provide a `#values` for params_batch to extract only values of objects from the params_batch
 - #363 - Too shallow ruby-kafka version lock